Extract AssignTypeTo impls to assign.go
so that the whole of the type-assignment operation is easier to
discover / navigate / understand.
